+++ R/calc_Tbreadth.R
@@ -36,7 +36,7 @@
 #'
 #' @examples
 #' example_file <- system.file(
-#'   "extdata", "Fish_8_13_3_example.csv",
+#'   "extdata", "Fish_7_13_2.csv",
 #'   package = "ShuttleboxR"
 #' )
 #' fish <- read_shuttlesoft(example_file)