+++ R/plot_coreT_histogram.R
@@ -25,7 +25,7 @@
 #'
 #' @examples
 #' example_file <- system.file(
-#'   "extdata", "Fish_8_13_3_example.csv",
+#'   "extdata", "Fish_7_13_2.csv",
 #'   package = "ShuttleboxR"
 #' )
 #' fish <- read_shuttlesoft(example_file)
@@ -135,7 +135,7 @@ plot_coreT_histogram <- function(data,
       x = "Core body temperature (°C)",
       y = "Percentage of observations (%)"
     ) +
-    ggplot2::theme_light()
+    ggplot2::theme_classic()
 
   if (isTRUE(show_Tpref)) {
     hist_plot <- hist_plot +